requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
dnspython==2.4.2
executing==1.2.0
filelock==3.12.2
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.42.1
idna==3.4
Imageio==2.8.0
ipykernel==6.25.1
ipython==7.30.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.21.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pymongo==4.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.7
